games/lab2/requirements.txt
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.2
kiwisolver==1.4.9
lxml==6.0.2
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
reportlab==4.4.3
scipy==1.16.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
